+++ pynenc/serializer/__init__.py
@@ -1,11 +1,12 @@
 from pynenc.serializer.base_serializer import BaseSerializer
 from pynenc.serializer.constants import ReservedKeys
 from pynenc.serializer.json_pickle_serializer import JsonPickleSerializer
-from pynenc.serializer.json_serializer import JsonSerializer
+from pynenc.serializer.json_serializer import JsonSerializable, JsonSerializer
 from pynenc.serializer.pickle_serializer import PickleSerializer
 
 __all__ = [
     "BaseSerializer",
+    "JsonSerializable",
     "JsonSerializer",
     "JsonPickleSerializer",
     "PickleSerializer",